+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/components/estimators/UniformTourEstimator.java
@@ -1,7 +1,7 @@
 package ch.ethz.matsim.discrete_mode_choice.components.estimators;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.matsim.api.core.v01.population.Person;
 
@@ -10,6 +10,8 @@ import ch.ethz.matsim.discrete_mode_choice.model.tour_based.DefaultTourCandidate
 import ch.ethz.matsim.discrete_mode_choice.model.tour_based.TourCandidate;
 import ch.ethz.matsim.discrete_mode_choice.model.tour_based.TourEstimator;
 import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.DefaultTripCandidate;
+import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.TripCandidate;
+import ch.ethz.matsim.discrete_mode_choice.replanning.time_interpreter.TimeInterpreter;
 
 /**
  * This estimator simply return a zero utility for every tour candidate that it
@@ -18,10 +20,34 @@ import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.DefaultTr
  * @author sebhoerl
  */
 public class UniformTourEstimator implements TourEstimator {
+	private final TimeInterpreter.Factory timeInterpreterFactory;
+
+	public UniformTourEstimator(TimeInterpreter.Factory timeInterpreterFactory) {
+		this.timeInterpreterFactory = timeInterpreterFactory;
+	}
+
 	@Override
 	public TourCandidate estimateTour(Person person, List<String> modes, List<DiscreteModeChoiceTrip> trips,
 			List<TourCandidate> previousTours) {
-		return new DefaultTourCandidate(1.0,
-				modes.stream().map(mode -> new DefaultTripCandidate(1.0, mode)).collect(Collectors.toList()));
+		List<TripCandidate> tripCandidates = new ArrayList<>(modes.size());
+
+		TimeInterpreter time = timeInterpreterFactory.createTimeInterpreter();
+		time.setTime(trips.get(0).getDepartureTime());
+
+		for (int index = 0; index < modes.size(); index++) {
+			DiscreteModeChoiceTrip trip = trips.get(index);
+
+			if (index > 0) { // We're already at the end of the first origin activity
+				time.addActivity(trip.getOriginActivity());
+				trip.setDepartureTime(time.getCurrentTime());
+			}
+
+			time.addPlanElements(trip.getInitialElements());
+
+			double duration = time.getCurrentTime() - trip.getDepartureTime();
+			tripCandidates.add(new DefaultTripCandidate(1.0, modes.get(index), duration));
+		}
+
+		return new DefaultTourCandidate(1.0, tripCandidates);
 	}
 }