+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/DiscreteModeChoiceTrip.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.matsim.api.core.v01.population.Activity;
 import org.matsim.api.core.v01.population.PlanElement;
+import org.matsim.core.utils.misc.Time;
 
 /**
  * This class represents an agent's trip. It contains structural information
@@ -16,17 +17,16 @@ public final class DiscreteModeChoiceTrip {
 	private final Activity originActivity;
 	private final Activity destinationActivity;
 	private final String initialMode;
-	private final double departureTime;
+	private double departureTime = Time.getUndefinedTime();
 	private final List<? extends PlanElement> initialElements;
 
 	private final int hashCode;
 
 	public DiscreteModeChoiceTrip(Activity originActivity, Activity destinationActivity, String initialMode,
-			List<? extends PlanElement> initialElements, double departureTime, int personHash, int tripHash) {
+			List<? extends PlanElement> initialElements, int personHash, int tripHash) {
 		this.originActivity = originActivity;
 		this.destinationActivity = destinationActivity;
 		this.initialMode = initialMode;
-		this.departureTime = departureTime;
 		this.initialElements = initialElements;
 
 		int hashCode = 12;
@@ -45,6 +45,10 @@ public final class DiscreteModeChoiceTrip {
 
 	public double getDepartureTime() {
 		return departureTime;
+	}
+
+	public void setDepartureTime(double departureTime) {
+		this.departureTime = departureTime;
 	}
 
 	public String getInitialMode() {