+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/trip_based/candidates/DefaultTripCandidate.java
@@ -8,10 +8,12 @@ package ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates;
 public class DefaultTripCandidate implements TripCandidate {
 	final private double utility;
 	final private String mode;
+	final private double duration;
 
-	public DefaultTripCandidate(double utility, String mode) {
+	public DefaultTripCandidate(double utility, String mode, double duration) {
 		this.utility = utility;
 		this.mode = mode;
+		this.duration = duration;
 	}
 
 	@Override
@@ -22,5 +24,10 @@ public class DefaultTripCandidate implements TripCandidate {
 	@Override
 	public String getMode() {
 		return mode;
+	}
+	
+	@Override
+	public double getDuration() {
+		return duration;
 	}
 }